refactor: move FromSchema to types file
