+++ packages/api/src/lockfileSchema.ts
@@ -1,4 +1,4 @@
-import type { FromSchema } from '@readme/api-core/lib';
+import type { FromSchema } from '@readme/api-core/types';
 
 import { SupportedLanguages } from './codegen/factory.js';
