+++ packages/core/src/lib/index.ts
@@ -1,9 +1,7 @@
-import type { FromSchema } from 'json-schema-to-ts';
-
 import getJSONSchemaDefaults from './getJSONSchemaDefaults.js';
 import parseResponse from './parseResponse.js';
 import prepareAuth from './prepareAuth.js';
 import prepareParams from './prepareParams.js';
 import prepareServer from './prepareServer.js';
 
-export { FromSchema, getJSONSchemaDefaults, parseResponse, prepareAuth, prepareParams, prepareServer };
+export { getJSONSchemaDefaults, parseResponse, prepareAuth, prepareParams, prepareServer };
