+++ packages/core/src/types.ts
@@ -1,3 +1,5 @@
+export { FromSchema } from 'json-schema-to-ts';
+
 export interface ConfigOptions {
   /**
    * Override the default `fetch` request timeout of 30 seconds. This number should be represented
