+++ packages/test-utils/sdks/response-title-quirks/src/types.ts
@@ -1,4 +1,4 @@
-import type { FromSchema } from '@readme/api-core/lib';
+import type { FromSchema } from '@readme/api-core/types';
 import * as schemas from './schemas';
 
 export type GetAnythingMetadataParam = FromSchema<typeof schemas.GetAnything.metadata>;
